Fixes gamerole to avoid params and readRole();
params and readRole(); have been bugging up recently due to changes to parameters.js. In this update I use message.content.substr(6); to exclude the "!gamerole " part of the command. This also allows me to retrieve the full content, which paraams(); was unable to do. Retrieving the full input let gameroles like clash of clans and clash royale (which both have the same first word) both be retrievable.